requirements.txt
Django==4.1.5
django-environ==0.9.0
django-storages==1.13
numpy==1.21.5
tqdm==4.64.1
praw==7.6.1
prawcore==2.3.0
requests==2.25.1
psycopg2-binary==2.9.4
pytz==2022.1
google-cloud-logging==3.5.0
google-cloud-secret-manager==2.12.0